chore: simplifying files and validation for password and email
