+++ app/routes/login/login.tsx
@@ -1,53 +1,11 @@
 import React, { useState } from "react";
-import { Layout } from "../components/layout";
+import { Layout } from "../../components/layout";
 import { Textfield } from "~/components/textField";
 import { Link, useActionData } from "@remix-run/react";
-import {
-  LoaderFunction,
-  ActionFunction,
-  json,
-  MetaFunction,
-} from "@remix-run/node";
-import { authUser } from "~/utils/user.server";
-
-import { authenticator } from "~/utils/auth.server";
+import { MetaFunction } from "@remix-run/node";
 
 export const meta: MetaFunction = () => {
   return [{ title: "New Remix App Login" }];
-};
-
-export const loader: LoaderFunction = async ({ request }) => {
-  const user = await authenticator.isAuthenticated(request, {
-    successRedirect: "/",
-  });
-  return user;
-};
-
-export const action: ActionFunction = async ({ request }) => {
-  const requestClone = request.clone(); // Cloning the request
-  const formData = await requestClone.formData();
-
-  const email = formData.get("email") as string;
-  const password = formData.get("password") as string;
-
-  const result = await authUser({ email, password });
-
-  if (!email || !password) {
-    return json(
-      { error: `All fields are required`, form: action },
-      { status: 400 }
-    );
-  }
-
-  if (!result.success) {
-    return json({ error: result.message }, { status: 400 });
-  }
-
-  return await authenticator.authenticate("form", request, {
-    successRedirect: "/",
-    failureRedirect: "/login",
-    context: { formData: formData },
-  });
 };
 
 interface ActionData {
@@ -117,3 +75,5 @@ export default function Login() {
     </Layout>
   );
 }
+
+export { loader, action } from "./login.server";
