+++ app/routes/signup/signup.tsx
@@ -1,55 +1,14 @@
 import React, { useState } from "react";
 import { Link, useActionData } from "@remix-run/react";
-import {
-  LoaderFunction,
-  ActionFunction,
-  json,
-  MetaFunction,
-} from "@remix-run/node";
+import { MetaFunction } from "@remix-run/node";
 
-import { authenticator } from "~/utils/auth.server";
-import { createUser } from "~/utils/user.server";
+import { Layout } from "../../components/layout";
+import { Textfield } from "../../components/textField";
 
-import { Layout } from "../components/layout";
-import { Textfield } from "~/components/textField";
+import { loader, action } from "./signup.server";
 
 export const meta: MetaFunction = () => {
   return [{ title: "New Remix App Signup" }];
-};
-
-export const loader: LoaderFunction = async ({ request }) => {
-  const user = await authenticator.isAuthenticated(request, {
-    successRedirect: "/",
-  });
-  return { user };
-};
-
-export const action: ActionFunction = async ({ request }) => {
-  const requestClone = request.clone(); // Cloning the request
-  const formData = await requestClone.formData();
-  const action = formData.get("_action") as string;
-  const email = formData.get("email") as string;
-  const password = formData.get("password") as string;
-  const name = formData.get("name") as string;
-
-  if (!email || !password || !name) {
-    return json(
-      { error: `All fields are required`, form: action },
-      { status: 400 }
-    );
-  }
-
-  const result = await createUser({ email, password, name });
-
-  if (!result.success) {
-    return json({ error: result.message }, { status: 400 });
-  }
-
-  return await authenticator.authenticate("form", request, {
-    successRedirect: "/",
-    failureRedirect: "/signup",
-    context: { formData: formData },
-  });
 };
 
 interface ActionData {
@@ -69,7 +28,6 @@ export default function SignUp() {
     name: actionData?.fields?.name || "",
   });
 
-  // Updates the form data when an input changes
   const handleInputChange = (
     event: React.ChangeEvent<HTMLInputElement>,
     field: string
@@ -129,3 +87,5 @@ export default function SignUp() {
     </Layout>
   );
 }
+
+export { loader, action };
