+++ app/utils/validators.ts
@@ -1,0 +1,29 @@
+function validatePassword(password: string): string | null {
+    if (password.length < 8) {
+        return "Password must be at least 8 characters long";
+    }
+    if (!/[A-Z]/.test(password)) {
+        return "Password must contain at least one uppercase letter";
+    }
+    if (!/[a-z]/.test(password)) {
+        return "Password must contain at least one lowercase letter";
+    }
+    if (!/[0-9]/.test(password)) {
+        return "Password must contain at least one number";
+    }
+    if (!/[!@#$%^&*(),.?":{}|<>]/.test(password)) {
+        return "Password must contain at least one special character";
+    }
+    return null;
+}
+
+function validateEmail(email: string): string | null {
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    if (!emailRegex.test(email)) {
+        return "Invalid email address";
+    }
+    return null; // No error, email is valid
+}
+
+
+export { validatePassword, validateEmail }
